// admin/src/api/generated/models/HostKeyScanResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * deploy-go-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface HostKeyScanResponse
 */
export interface HostKeyScanResponse {
    /**
     *
     * @type {string}
     * @memberof HostKeyScanResponse
     */
    checkId: string;
    /**
     *
     * @type {string}
     * @memberof HostKeyScanResponse
     */
    fingerprint: string;
    /**
     *
     * @type {string}
     * @memberof HostKeyScanResponse
     */
    snapshotHash: string;
}

/**
 * Check if a given object implements the HostKeyScanResponse interface.
 */
export function instanceOfHostKeyScanResponse(value: object): value is HostKeyScanResponse {
    if ((!('checkId' in (value as Record<string, any>)) && !('check_id' in (value as Record<string, any>))) || ((value as Record<string, any>)['checkId'] === undefined && (value as Record<string, any>)['check_id'] === undefined)) return false;
    if (!('fingerprint' in value) || value['fingerprint'] === undefined) return false;
    if ((!('snapshotHash' in (value as Record<string, any>)) && !('snapshot_hash' in (value as Record<string, any>))) || ((value as Record<string, any>)['snapshotHash'] === undefined && (value as Record<string, any>)['snapshot_hash'] === undefined)) return false;
    return true;
}

export function HostKeyScanResponseFromJSON(json: any): HostKeyScanResponse {
    return HostKeyScanResponseFromJSONTyped(json, false);
}

export function HostKeyScanResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): HostKeyScanResponse {
    if (json == null) {
        return json;
    }
    return {

        'checkId': json['check_id'],
        'fingerprint': json['fingerprint'],
        'snapshotHash': json['snapshot_hash'],
    };
}

export function HostKeyScanResponseToJSON(json: any): HostKeyScanResponse {
    return HostKeyScanResponseToJSONTyped(json, false);
}

export function HostKeyScanResponseToJSONTyped(value?: HostKeyScanResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'check_id': value['checkId'],
        'fingerprint': value['fingerprint'],
        'snapshot_hash': value['snapshotHash'],
    };
}