// admin/src/api/generated/models/NodeCheckResponse.ts
export interface NodeCheckResponse {
    /**
     *
     * @type {string}
     * @memberof NodeCheckResponse
     */
    architecture?: string | null;
    /**
     *
     * @type {string}
     * @memberof NodeCheckResponse
     */
    createdAt: string;
    /**
     *
     * @type {number}
     * @memberof NodeCheckResponse
     */
    diskAvailableBytes?: number | null;
    /**
     *
     * @type {string}
     * @memberof NodeCheckResponse
     */
    failureCode?: string | null;
    /**
     *
     * @type {string}
     * @memberof NodeCheckResponse
     */
    failureMessage?: string | null;
    /**
     *
     * @type {string}
     * @memberof NodeCheckResponse
     */
    finishedAt?: string | null;
    /**
     *
     * @type {string}
     * @memberof NodeCheckResponse
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof NodeCheckResponse
     */
    osName?: string | null;
    /**
     *
     * @type {string}
     * @memberof NodeCheckResponse
     */
    status: string;
}